+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.2 | 28.07.2020 <html broser name, classes update>
+* add browser name as css class to html
+* imgDC class version 2.1.1
+* template class version 2.3.3
+* WPseo class version 2.1.3
+
 ## 1.1 | 28.07.2020 <template header, classes update>
 * add frontend css class to body
 * WPinit class version 2.5.2